login: added registration for initial setup
If you're not running in debug mode you won't have a user set up when you
first install the app; now if you try to use the app without a user you'll
be redirected to a registration page to set up your initial user.

I should probably have this as part of a nice setup script that'll write
your config file like WP does, but that'll come later.